Move the api test jest config into a separate file.
Move the api config into a separate file to allow easier testing.
Testing code still needs removing.
Tear down the web server between test files.
