+++ package.json
@@ -85,7 +85,7 @@
     "test": "craco test --testPathPattern=src/web/",
     "api": "nodemon -r dotenv/config ./src/api/api.ts",
     "build-api": "npx tsc --build ./tsconfig-api.json",
-    "test-api": "npx jest --testTimeout=5000 --testPathPattern=src/api/tests --setupFiles dotenv/config --forceExit",
+    "test-api": "npx jest --testTimeout=5000 --config=./src/api/jest.config.ts --forceExit",
     "storybook": "start-storybook -p 6006 -s public",
     "build-storybook": "build-storybook -s public",
     "stylelint-check": "stylelint-config-prettier-check",
@@ -180,6 +180,7 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-simple-import-sort": "^8.0.0",
     "eslint-plugin-testing-library": "^5.9.1",
+    "http-terminator": "^3.2.0",
     "jest": "^29.5.0",
     "keycloak-request-token": "^0.1.0",
     "nodemon": "^2.0.20",
