+++ src/api/jest.config.ts
@@ -1,0 +1,9 @@
+import type { Config } from 'jest';
+
+const config: Config = {
+  rootDir: '../../',
+  setupFiles: ['dotenv/config', './src/api/jest.setup.ts'],
+  testMatch: ['**/api/tests/*.test.ts', '**/api/tests/*.spec.ts'],
+};
+
+export default config;
