+++ src/api/jest.setup.ts
@@ -1,0 +1,4 @@
+import { Configure } from '../database/SelfServeDatabase';
+
+console.log('CONFIGURING DATABASE FOR TESTS..............');
+Configure();
